InstaRound

Documentation for InstaRound.

InstaRound.IGRound

Base.round

InstaRound.business_round

InstaRound.IGRound — Type

IGRound

Abstract base type for dispatched InstaRound rounds.

Base.round — Method

round(::Type{IGRound}, x::Number)

Main function to round a number x to the nearest IG style.

Arguments

x: The number to round.

# Example
julia> round(IGRound, 100_000)
"100K"

InstaRound.business_round — Method

business_round(num)

Internal function to round a number to the nearest IG style.